Refactor code base in auth
