+++ core/libs/data/auth.ts
@@ -2,7 +2,7 @@ import useSWR, { Fetcher } from 'swr'
 import { UserClaim } from '../auth'
 
 const fetchUserHandler: Fetcher<UserClaim, string> = async () => {
-  return fetch('/api/me', { credentials: 'include' }).then((res) => res.json())
+  return fetch('/api/auth/me', { credentials: 'include' }).then((res) => res.json())
 }
 
 export default function useAuth() {
@@ -10,7 +10,7 @@ export default function useAuth() {
     data: fetchedUser,
     mutate,
     error,
-  } = useSWR('/api/me', fetchUserHandler)
+  } = useSWR('/api/auth/me', fetchUserHandler)
   const loading = !fetchedUser && !error
   const loggedOut = !!error && (error.status === 401 || error.status === 500)
   return { fetchedUser, mutate, loading, loggedOut }
